RISC-V: Add experimental support for Zvabd extension
This patch adds experimental support for the RISC-V Zvabd (Vector Absolute
Difference) extension, based on version 0.5 of the proposal.

The Zvabd instruction set specification can be found in the fast-track proposal:
https://lf-riscv.atlassian.net/wiki/spaces/VXXX/pages/166690866/Fast-Track+Proposal+for+integer+vector+absolute+difference+instructions

Note: This patch is just an initial prototype, mainly for early evaluation (e.g.
on FPGA) and to support fast-track proposal progress; it is not meant for
merging yet.

Co-authored-by: Yuke Tang <tang.yuke@zte.com.cn>
Signed-off-by: Zhongyao Chen <chen.zhongyao@zte.com.cn>